# apps/rest-api/docs/swagger/vector_api.yaml
openapi: 3.0.0
info:
  title: Vector API
  description: API endpoints for vector embedding operations
  version: 1.0.0

tags:
  - name: vectors
    description: Vector embedding operations

paths:
  /api/v1/vectors/store:
    post:
      summary: Store a vector embedding
      tags:
        - vectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Successfully stored embedding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Embedding'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/vectors/search:
    post:
      summary: Search vector embeddings
      description: Search embeddings by similarity to a query vector with filtering options
      tags:
        - vectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/vectors/context/{contextId}:
    get:
      summary: Get all embeddings for a context
      tags:
        - vectors
      parameters:
        - name: contextId
          in: path
          description: Context ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Context embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeddings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Embedding'
        '404':
          description: Context not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete all embeddings for a context
      tags:
        - vectors
      parameters:
        - name: contextId
          in: path
          description: Context ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/vectors/models:
    get:
      summary: Get supported model IDs
      description: Returns list of all model IDs with stored embeddings
      tags:
        - vectors
      responses:
        '200':
          description: List of model IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/vectors/context/{contextId}/model/{modelId}:
    get:
      summary: Get embeddings for a specific model in a context
      tags:
        - vectors
      parameters:
        - name: contextId
          in: path
          description: Context ID
          required: true
          schema:
            type: string
        - name: modelId
          in: path
          description: Model ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeddings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Embedding'
        '404':
          description: Context or model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete all embeddings for a specific model in a context
      description: Delete embeddings for a specific model in a context. This is a new feature that allows for fine-grained control over which embeddings to delete based on model ID.
      tags:
        - vectors
      parameters:
        - name: contextId
          in: path
          description: Context ID
          required: true
          schema:
            type: string
        - name: modelId
          in: path
          description: Model ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Embedding:
      type: object
      properties:
        id:
          type: string
        context_id:
          type: string
          description: In the API, this maps to tenant_id through the adapter pattern
        content_index:
          type: integer
        text:
          type: string
          description: In the API, this maps to content field through the adapter pattern
        embedding:
          type: array
          items:
            type: number
            format: float
        model_id:
          type: string
        metadata:
          type: object
          additionalProperties: true

    EmbeddingRequest:
      type: object
      required:
        - context_id
        - content_index
        - text
        - embedding
        - model_id
      properties:
        context_id:
          type: string
          description: Context identifier (maps to tenant_id in API models)
        content_index:
          type: integer
          description: Index of the content within the context
        text:
          type: string
          description: Text content to embed (maps to content in API models)
        embedding:
          type: array
          description: Vector embedding values
          items:
            type: number
            format: float
        model_id:
          type: string
          description: Identifier of the model used for embedding
        metadata:
          type: object
          description: Additional metadata
          additionalProperties: true

    SearchRequest:
      type: object
      required:
        - query_embedding
        - context_id
      properties:
        query_embedding:
          type: array
          description: Query vector embedding
          items:
            type: number
            format: float
        context_id:
          type: string
          description: Context to search within (maps to tenant_id in API models)
        model_id:
          type: string
          description: Optional model ID filter
        limit:
          type: integer
          description: Maximum number of results
          default: 10
        similarity_threshold:
          type: number
          format: float
          description: Minimum similarity score (0-1)
          default: 0.7

    SearchResults:
      type: object
      properties:
        results:
          type: array
          description: Search result embeddings
          items:
            $ref: '#/components/schemas/SearchResult'

    SearchResult:
      type: object
      properties:
        embedding:
          $ref: '#/components/schemas/Embedding'
        score:
          type: number
          format: float
          description: Similarity score (0-1)

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code